requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
black==24.8.0
click==8.1.7
frozenlist==1.4.1
furl==2.1.3
idna==3.10
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
multidict==6.1.0
mypy-extensions==1.0.0
orderedmultidict==1.0.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dotenv==1.0.1
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
typing_extensions==4.12.2
xmltodict==0.13.0
yarl==1.11.1
